+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Full event editor support for `quest`, `shop`, `arena` and `trade` events
+- Ability to view and edit `OPEN_QUEST_DIALOG` event branches
 - Ability to view and edit `START_NPC_TRADE_MENU` event branches
 
 ### Fixed